requirements.txt
discord.py
import-expression
ciberedev